refactor(ios): split foreground geolocation helpers
